third_party/blink/web_tests/platform/win11-arm64/http/tests/misc/async-script-removed-expected.txt

CONSOLE ERROR: Uncaught SyntaxError: Unexpected token '<'
CONSOLE ERROR: Uncaught SyntaxError: Unexpected token '<'
Test that a synchronous script tag is executed in the right order, even if it's removed from the document before it was loaded.

This test passes if the number "4123" is displayed below.

43
